split fantasy plugin into multiple files
